Add new verifier class and refactor vrf related data and functions
